+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * 2.5.5
     - added switch button for count
     - fixed UI performance issue
+    - fixed sourcemap content issue
 
 * 2.5.4
     - fixed empty coverage